created validations
