+++ trading/src/main/java/com/cryptocurrency/trading/Controllers/UserController.java
@@ -3,8 +3,11 @@ package com.cryptocurrency.trading.Controllers;
 import com.cryptocurrency.trading.Models.User;
 import com.cryptocurrency.trading.Service.AccountResetService;
 import com.cryptocurrency.trading.Service.UserService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,6 +18,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.math.BigDecimal;
 import java.sql.SQLException;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/user")
@@ -25,12 +30,25 @@ public class UserController {
 
 
     @PostMapping("/create")
-    public ResponseEntity<?> createUser(@RequestBody User user) {
+    public ResponseEntity<?> createUser(@Valid @RequestBody User user, BindingResult bindingResult) throws SQLException {
+        userService.userValidator(user, bindingResult);
+
+        if (bindingResult.hasErrors()) {
+            Map<String, String> errors = bindingResult.getFieldErrors()
+                    .stream()
+                    .collect(Collectors.toMap(
+                            FieldError::getField,
+                            fieldError -> fieldError.getDefaultMessage(),
+                            (existing, replacement) -> existing // Ако има повече от една грешка за дадено поле, вземи първата
+                    ));
+            return ResponseEntity.badRequest().body(Map.of("errors", errors));
+        }
+
         try {
-            userService.createUser(user);
-            return ResponseEntity.ok("User is created successfully");
+            userService.createUser(user, bindingResult);
+            return ResponseEntity.ok(Map.of("message", "User is created successfully"));
         } catch (IllegalArgumentException e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
+            return ResponseEntity.badRequest().body(Map.of("errors", "username:" + e.getMessage()));
         } catch (SQLException e) {
             return ResponseEntity.internalServerError().body("Възникна грешка с базата данни");
         }
