+++ trading/src/main/java/com/cryptocurrency/trading/Models/User.java
@@ -1,18 +1,25 @@
 package com.cryptocurrency.trading.Models;
 
+import jakarta.validation.constraints.DecimalMin;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.math.BigDecimal;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Setter
 @Getter
 public class User {
     private int id;
+    @NotBlank(message = "Username is required")
+    @Size(min = 3, max = 20, message = "Username must be between 3 and 20 characters")
     private String username;
+
+    @DecimalMin(value = "0.0", message = "Balance must be positive")
     private BigDecimal balance;
 }
