+++ trading/src/main/java/com/cryptocurrency/trading/Service/UserAssetService.java
@@ -1,53 +1,25 @@
 package com.cryptocurrency.trading.Service;
 
-import com.cryptocurrency.trading.DAO.UserAssetDao;
 import com.cryptocurrency.trading.Models.UserAsset;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.sql.SQLException;
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-public class UserAssetService {
+public interface UserAssetService {
+    List<UserAsset> getAssetsByUserId(int userId) throws SQLException;
 
+    boolean userAssetExists(int userId, String symbol) throws SQLException;
 
-    private final UserAssetDao userAssetDao;
+    void updateUserAssetQuantity(int userId, String symbol, BigDecimal quantityToAdd) throws SQLException;
 
-    public List<UserAsset> getAssetsByUserId(int userId) throws SQLException {
-        return userAssetDao.findByUserId(userId);
-    }
+    void insertUserAsset(UserAsset userAsset) throws SQLException;
 
+    BigDecimal getCryptoQuantity(int userId, String symbol) throws SQLException;
 
-    public boolean userAssetExists(int userId, String symbol) throws SQLException {
-        return userAssetDao.userAssetExists(userId, symbol);
-    }
+    void reduceQuantityBySymbol(int userId, String symbol, BigDecimal quantityToRemove) throws SQLException;
 
-    public void updateUserAssetQuantity(int userId, String symbol, BigDecimal quantityToAdd) throws SQLException {
-        userAssetDao.updateUserAssetQuantity(userId, symbol, quantityToAdd);
-    }
+    void remove(int userId, String symbol) throws SQLException;
 
-    public void insertUserAsset(UserAsset userAsset) throws SQLException {
-        userAssetDao.insertUserAsset(userAsset);
-    }
-
-
-    public BigDecimal getCryptoQuantity(int userId, String symbol) throws SQLException {
-        return userAssetDao.findCryptoQuantity(userId, symbol);
-    }
-
-    public void reduceQuantityBySymbol(int userId, String symbol, BigDecimal quantityToRemove) throws SQLException {
-        userAssetDao.reduceQuantityBySymbol(userId, symbol, quantityToRemove);
-    }
-
-
-    public void remove(int userId, String symbol) throws SQLException {
-        userAssetDao.remove(userId, symbol);
-    }
-
-    public void removeUserAssetsByUserId(int userId) throws Exception {
-        userAssetDao.removeUserAssetsByUserId(userId);
-    }
+    void removeUserAssetsByUserId(int userId) throws Exception;
 }
