+++ trading/src/main/java/com/cryptocurrency/trading/Service/UserService.java
@@ -1,46 +1,23 @@
 package com.cryptocurrency.trading.Service;
 
-import com.cryptocurrency.trading.DAO.UserDao;
 import com.cryptocurrency.trading.Models.User;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
+import org.springframework.validation.BindingResult;
 
 import java.math.BigDecimal;
 import java.sql.SQLException;
 
-@Service
-@RequiredArgsConstructor
-public class UserService {
+public interface UserService {
+    void createUser(User user, BindingResult bindingResult) throws SQLException;
 
-    private final UserDao userDao;
-    private final UserAssetService userAssetService;
-    /*private final TransactionService transactionService;*/
+    void userValidator(User user, BindingResult bindingResult) throws SQLException;
 
+    User getUserByUsername(String username) throws SQLException;
 
-    public void createUser(User user) throws SQLException {
-        if (userDao.userExists(user.getUsername())) {
-            throw new IllegalArgumentException("Потребител с това име вече съществува!");
-        }
-        userDao.create(user);
-    }
+    void updateBalance(int userId, BigDecimal newBalance) throws SQLException;
 
-    public User getUserByUsername(String username) throws SQLException {
-        return userDao.findByUsername(username);
-    }
+    boolean hasEnoughAndSubtractBalance(int userId, BigDecimal amount) throws SQLException;
 
-    public void updateBalance(int userId, BigDecimal newBalance) throws SQLException {
-        userDao.updateBalance(userId, newBalance);
-    }
+    void resetBalance(int userId) throws Exception;
 
-    public boolean hasEnoughAndSubtractBalance(int userId, BigDecimal amount) throws SQLException {
-        return userDao.hasEnoughAndSubtractBalance(userId, amount);
-    }
-
-    public void resetBalance(int userId) throws Exception {
-        userDao.resetBalance(userId);
-    }
-
-    public BigDecimal getBalance(int userId) throws SQLException {
-        return userDao.findBalance(userId);
-    }
+    BigDecimal getBalance(int userId) throws SQLException;
 }
